Added API, moved parsing options to utils
